feat(playerbot): Migrate to WoW 12.0.0 compatibility
Merge upstream TrinityCore 12.0.0 and update Playerbot module:

## API Changes
- Difficulty enum: uint8 → int16 (PlayerBotHooks.h, GroupCoordinator.h)
- CastItemUseSpell: int32[2] → std::array<int32, 3> (7 files, 10 call sites)

## Packet Protocol Updates (SpellPacketBuilder.cpp)
- SpellTargetData: Unknown1127_1 → HousingGUID, Unknown1127_2 → HousingIsResident
- SpellCastRequest: Added Misc[2], renamed OptionalCurrencies → ExtraCurrencyCosts,
  OptionalReagents → CraftingReagents, RemovedModifications → RemovedReagents,
  CraftingFlags → CraftingCastFlags

## Stats System (STAT_SPIRIT re-added, MAX_STATS = 5)
- LFGRoleDetector: Added spirit to healer scoring
- RoleAssignment: Added spirit to healer role modifier
- DoubleBufferedSpatialGrid: Added spirit field to PlayerSnapshot

+++ src/modules/Playerbot/Advanced/GroupCoordinator.h
@@ -23,7 +23,7 @@ class Player;
 class Group;
 class Unit;
 class Creature;
-enum Difficulty : uint8;
+enum Difficulty : int16;
 
 namespace Playerbot
 {
